[Taranis] one-click model select
When on the model select screen, pressing [enter] loads the model currently select, instead of having to press enter long + click “select model” in the pop up menu.

It might not be the case for everyone, but it seems to me the most intuitive action to execute, and should thus not require the user to trigger the pop-up menu.